add new skeleton
