+++ jobs/Backend/Task/src/Application/Providers/CzechNationalBank/CzechNationalBankHttpClient.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+
+/// <summary>
+/// HTTP client for fetching Czech National Bank exchange rate data.
+/// </summary>
+public class CzechNationalBankHttpClient
+{
+    public async Task<string> GetDailyRatesAsync(string url, CancellationToken cancellationToken)
+    {
+        throw new NotImplementedException();
+    }
+}
