+++ jobs/Backend/Task/src/Domain/Currency.cs
@@ -1,0 +1,19 @@
+﻿namespace ExchangeRateUpdater.Domain;
+
+public class Currency
+{
+    public Currency(string code)
+    {
+        Code = code;
+    }
+
+    /// <summary>
+    /// Three-letter ISO 4217 code of the currency.
+    /// </summary>
+    public string Code { get; }
+
+    public override string ToString()
+    {
+        return Code;
+    }
+}
