+++ jobs/Backend/Task/src/Domain/ExchangeRate.cs
@@ -1,0 +1,22 @@
+﻿namespace ExchangeRateUpdater.Domain;
+
+public class ExchangeRate
+{
+    public ExchangeRate(Currency sourceCurrency, Currency targetCurrency, decimal value)
+    {
+        SourceCurrency = sourceCurrency;
+        TargetCurrency = targetCurrency;
+        Value = value;
+    }
+
+    public Currency SourceCurrency { get; }
+
+    public Currency TargetCurrency { get; }
+
+    public decimal Value { get; }
+
+    public override string ToString()
+    {
+        return $"{SourceCurrency}/{TargetCurrency} = {Value:F4}";
+    }
+}
