+++ jobs/Backend/Task/src/Domain/IExchangeRateProvider.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace ExchangeRateUpdater.Domain;
+
+public interface IExchangeRateProvider
+{
+    IEnumerable<ExchangeRate> GetExchangeRates(IEnumerable<Currency> currencies);
+}
