+++ jobs/Backend/Task/tests/Application.UnitTests/UnitTest1.cs
@@ -1,0 +1,10 @@
+﻿namespace ExchangeRateUpdater.UnitTests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+
+    }
+}
